requirements.txt
streamlit
langchain_openai
pinecone
tavily
langgraph
pandas
numpy
python-dotenv
